% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eepd_boot.R
\name{eepd_boot}
\alias{eepd_boot}
\title{Bootstrap estimation of ATTs}
\usage{
eepd_boot(
  models,
  data,
  nboot = 999,
  boot_type = getOption("fwb_wtype", "exp"),
  nsim = 200,
  group_var,
  unit_var,
  time_var,
  val_times,
  post_time,
  cl = NULL,
  verbose = FALSE
)
}
\arguments{
\item{models}{either an \code{eepd_models} object (the output of a call to \code{\link[=eepd_mod]{eepd_mod()}}) or an \code{eepd_fits} object (the output of a call to \code{\link[=eepd_fit]{eepd_fit()}}). If the latter, the arguments \code{data}, \code{group_var}, \code{unit_var}, \code{time_var}, \code{val_times}, and \code{post_time} should be left empty as they will be extracted from the supplied object.}

\item{data}{a dataset containing all the variables named in the supplied models (i.e., the outcome and any predictors) as well as any variable named below.}

\item{nboot}{the number of bootstrap iterations to use; default is 999. More is better but takes longer.}

\item{boot_type}{string; the type of bootstrap to perform. See the \code{wtype} argument of \code{\link[fwb:fwb]{fwb::fwb()}} for allowable options. The default is \code{"exp"}, which requests the fractional weighted bootstrap using weights drawn from an Exp(1) distribution. \code{"multinom"} requests the usual bootstrap, which can fail when key observation requird to fit certain models happen not to be selected into a given bootstrap sample.}

\item{nsim}{the number of simulation iteration to perform in each bootstrap sample. Default is 200. More is better but takes longer.}

\item{group_var}{string; the name of the treatment variable in \code{data} defining the "to be treated" and "not to be treated" groups. The corresponding variable should take on values of 0 and 1 only.}

\item{unit_var}{string, optional; the name of the unit ID variable in \code{data} when any models suppiled to \code{models} contain unit fixed effects.}

\item{time_var}{string; the name of the variable in \code{data} containing the time variable.}

\item{val_times}{a numeric vector corresponding to the pre-treatment times that will be used as validation times when select the model with the optimal average expected prediction error.}

\item{post_time}{a number corresponding to the first post-treatment time period.}

\item{cl}{a cluster object created by \ifelse{\Sexpr[results=rd,stage=render]{length(find.package("parallel", quiet = TRUE, verbose = FALSE)) > 0}}{\code{\link[parallel:makeCluster]{parallel::makeCluster()}}}{\code{parallel::makeCluster()}}, or an integer to indicate the number of child-processes (integer values are ignored on Windows) for parallel evaluations. See \ifelse{\Sexpr[results=rd,stage=render]{length(find.package("pbapply", quiet = TRUE, verbose = FALSE)) > 0}}{\code{\link[pbapply:pblapply]{pbapply::pblapply()}}}{\code{pbapply::pblapply()}} for details. If \code{NULL}, no parallelization will take place.}

\item{verbose}{\code{logical}; whether to display a progress bar.}
}
\value{
An \code{fwb} object containing the estimated ATTs in each bootstrap iteration. See \code{\link[fwb:fwb]{fwb::fwb()}} for details. \code{summary()}, \code{plot()}, and \code{print()} methods are available; see \code{\link[fwb:summary.fwb]{fwb::summary.fwb()}} and \code{\link[fwb:plot.fwb]{fwb::plot.fwb()}} for details.
}
\description{
\code{eepd_boot()} bootstraps the selection of optimal models and estimation of ATTs done by \code{\link[=eepd_sim]{eepd_sim()}} in order to account for uncertainty in sampling from the population. Bootstrapping is done by \code{\link[fwb:fwb]{fwb::fwb()}}, which uses the fractional weighted bootstrap or the traditional bootstrap.
}
\examples{
data("ptpdata")

# Combination of 8 models: 2 baseline formulas,
# 2 families, 2 lags
models <- eepd_mod(list(crude_rate ~ 1,
                        crude_rate ~ year),
                   log = c(FALSE, TRUE))
models

# Fit the models to data; unit_var must be supplied for
# fixed effects
cl <- parallel::detectCores()
boot_out <- eepd_boot(models, data = ptpdata,
                      nboot = 99, nsim = 100,
                      group_var = "group",
                      time_var = "year",
                      val_times = 1999:2003,
                      post_time = 2008,
                      unit_var = "state",
                      cl = cl, verbose = TRUE)

summary(boot_out, ci.type = "perc")
}
\seealso{
\code{\link[=eepd_fit]{eepd_fit()}}; \code{\link[=eepd_sim]{eepd_sim()}}; \code{\link[fwb:fwb]{fwb::fwb()}}
}
