% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eepd_fit.R
\name{eepd_fit}
\alias{eepd_fit}
\title{Fit models to datasets}
\usage{
eepd_fit(
  models,
  data,
  weights = NULL,
  group_var,
  time_var,
  val_times,
  post_time,
  unit_var
)
}
\arguments{
\item{models}{an \code{eepd_models} object; the output of a call to \code{\link[=eepd_mod]{eepd_mod()}}.}

\item{data}{a dataset containing all the variables named in the supplied models (i.e., the outcome and any predictors) as well as any variable named below.}

\item{weights}{an optional vector of weights (e.g., sampling weights) used to fit weighted regression models.}

\item{group_var}{string; the name of the treatment variable in \code{data} defining the "to be treated" and "not to be treated" groups. The corresponding variable should take on values of 0 and 1 only.}

\item{time_var}{string; the name of the variable in \code{data} containing the time variable.}

\item{val_times}{a numeric vector corresponding to the pre-treatment times that will be used as validation times when select the model with the optimal average expected prediction error.}

\item{post_time}{a number corresponding to the first post-treatment time period.}

\item{unit_var}{string, optional; the name of the unit ID variable in \code{data} when any models suppiled to \code{models} contain unit fixed effects.}
}
\value{
An \code{eepd_fits} object, which is a list containing the models supplied to \code{models}, a grid of all fitted models, a list of all model fit objects, a list of all estimated coefficients, the joint covariance of the coefficients, the dataset supplied to \code{data}, and other components supplied to \code{eepd_fit()}.
}
\description{
\code{eepd_fit()} fits models to the pre-treatment and post-treatment models to obtain the joint distribution of model coefficients, which used by \code{\link[=eepd_sim]{eepd_sim()}} to compute the average differential prediction errors and ATTs.
}
\details{
\code{eepd_fit()} create a grid of all models and all time points and fits all corresponding models. For each validation time supplied to \code{val_times} and for the post-treatment time supplied to \code{post_time}, each model is fit using all previous times. For example, for a validation time of 5, a model is fit with data only from periods 1-4.

\code{\link[=glm]{glm()}} or \code{\link[MASS:glm.nb]{MASS::glm.nb()}} are used to fit the given models. The joint covariance matrix of all the coefficients is computed using the SUEST method described in Mize et al. (2019, p164), which is also used by the STATA command \code{suest}. This is equivalent to the covariance matrix computed by stacking the score equations for the models and fitting them using M-estimation and yields the equivalent of the HC0 covariance matrix for all within-model covariances.
}
\examples{
data("ptpdata")

# Combination of 8 models: 2 baseline formulas,
# 2 families, 2 lags
models <- eepd_mod(list(crude_rate ~ 1,
                         crude_rate ~ year),
                    family = list("gaussian", "quasipoisson"),
                    lag = 0:1, fixef = TRUE)
models

# Fit the models to data; unit_var must be supplied for
# fixed effects
fits <- eepd_fit(models, data = ptpdata,
                 group_var = "group",
                 time_var = "year",
                 val_times = 1999:2007,
                 post_time = 2008,
                 unit_var = "state")
}
\seealso{
\code{\link[=glm]{glm()}}; \code{\link[MASS:glm.nb]{MASS::glm.nb()}}; \code{\link[=eepd_sim]{eepd_sim()}} to compute the ATT and average absolute differential prediction errors; \code{\link[=eepd_boot]{eepd_boot()}} to bootstrap the models.
}
