% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eepd_mod.R
\name{eepd_mod}
\alias{eepd_mod}
\title{Generate models used to fit outcomes}
\usage{
eepd_mod(
  formula_list,
  family = "gaussian",
  lag = 0,
  diff_k = 0,
  log = FALSE,
  fixef = FALSE,
  identiy_only_log = TRUE
)
}
\arguments{
\item{formula_list}{a list of model formulas with the outcome on the left side and predictions (or just an intercept) on the right side.}

\item{family}{a list of family specifications; see \code{\link[=family]{family()}} for allowable options. These iwll eventually be passed to \code{\link[=glm]{glm()}} when fitting the models in \code{\link[=eepd_fit]{eepd_fit()}}, \code{\link[=eepd_sim]{eepd_sim()}}, or \code{\link[=eepd_boot]{eepd_boot()}}. \code{"negbin"} can also be supplied to request a negative binomial model fit using \code{\link[MASS:glm.nb]{MASS::glm.nb()}}. Default is \code{"gaussian"} to specify a linear model.}

\item{lag}{a vector of integers indicating the desired outcome lags to be used as predictors. For example, a \code{lag} value of 3 means the outcome lagged once, twice, and three times will be included as predictors. Default is 0 for no lags.}

\item{diff_k}{a vector of integers indicating the desired outcome lag to be used a an offset For example, a \code{diff_k} value of 3 means the third lag of the outcome will be included as an offset, equivalent to using the outcome minus its corresponding lag as the outcome of the corresponding model. Default is 0 for no lags. Any models with a \code{diff_k} value less than a \code{lag} value will be removed automatically.}

\item{log}{a logical vector indicating whether the outcome should be logged. Default is \code{FALSE} to use the original outcome.}

\item{fixef}{a logical vector indicating whether unit fixed effects should be included as predictors. Default is \code{FALSE} to omit unit fixed effects.}

\item{identiy_only_log}{\code{logical}; whether to omit any models in which \code{log} is \code{TRUE} but the link in the \code{family} specification corresponds to something other than \code{"identity"}. Default is \code{TRUE}, and this should probably not be changed.}
}
\value{
An \code{eepd_models} object, which is a list containing the full cross (less any omitted combinations) of the model features specified in the arguments. These have a \code{print()} method and can be combined using \code{c()}.
}
\description{
\code{eepd_mod()} generates a list of models characterized by a basic model formulas and other options (e.g., lags, families, etc.) that are supplied to \code{\link[=eepd_fit]{eepd_fit()}}, \code{\link[=eepd_sim]{eepd_sim()}}, or \code{\link[=eepd_boot]{eepd_boot()}}. These values are completely crossed to create a grid of model specifications, and multiple sets of model specifications can be combined using \code{c()} (see Examples).
}
\examples{
data("ptpdata")

# Combination of 8 models: 2 baseline formulas,
# 2 families, 2 lags
models1 <- eepd_mod(list(crude_rate ~ 1,
                         crude_rate ~ year),
                    family = list("gaussian", "quasipoisson"),
                    lag = 0:1, fixef = TRUE)
models1

# Add a single other model with a square time trend
models2 <- eepd_mod(crude_rate ~ poly(year, 2),
                    family = "gaussian",
                    fixef = FALSE)
models2

(models <- c(models1, models2))

# Remove a model
models[[4]] <- NULL
models
}
\seealso{
\link{formula}, \link{family}
}
