% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eepd_sim.R
\name{eepd_sim}
\alias{eepd_sim}
\alias{plot.eepd_sim}
\title{Estimate ATTS from models fits}
\usage{
eepd_sim(fits, nsim = 0, cl = NULL, verbose = FALSE)

\method{plot}{eepd_sim}(x, stack = TRUE, palette = "YlGnBu", ncol = NULL, ...)
}
\arguments{
\item{fits}{an \code{eepd_fits} object; the output of a call to \code{\link[=eepd_fit]{eepd_fit()}}.}

\item{nsim}{the number of simulation iterations to run. Set to 0 (the default) to use the original model coefficients instead of simulating. See Details.}

\item{cl}{a cluster specification to allow for parallel computing. Passed to the \code{cl} argument of \code{\link[pbapply:pbapply]{pbapply::pblapply()}}. Default is to use sequential evaluation on a single core.}

\item{verbose}{\code{logical}; when \code{nsim} is greater than 0, whether to print a progress bar. Default is \code{FALSE} to omit a progress bar.}

\item{x}{an \code{eepd_sim} object; the output of a call to \code{eepd_sim()}.}

\item{stack}{\code{logical}; whether to produced a stacked density plot of the ATTs or multiple overlapping density plots when \code{nsim} is greater than 0.}

\item{palette}{string; the color palette to use when \code{nsim} is greater than 0. Passed to \code{\link[ggplot2:scale_brewer]{ggplot2::scale_fill_brewer()}}.}

\item{ncol}{the number of columns in which to display the average absolute prediction errors when \code{nsim = 0}.}

\item{\dots}{ignored.}
}
\value{
An \code{eepd_sim} object, which contains the absolute prediction errors for models in all validation periods in each simulation, the simulated ATTs for all models in each simulation, the optimal model in each simulation, and the chosen ATT in each simulation. When \code{nsim = 0}, these correspond to the  values computed using the original model coefficients.

\code{plot()} returns a \code{ggplot} object when \code{nsim = 0} and a \code{patchwork} object otherwise.

The \code{print()} method either displays the ATT and the model used to compute it (when \code{nsim = 0}) or the average of the simulated ATTs and a frequency table of the chosen models.
}
\description{
\code{eepd_sim()} computes the ATTs from the models previously fit by \code{\link[=eepd_fit]{eepd_fit()}}, choosing the optimal one by minimizing the largest absolute average prediction error across validation times. Optionally, this process can be simulated to arrive at a distribution of ATTs that accounts for the uncertainty in selecting the optimal model. \code{plot()} plotst he resulting ATT(s).
}
\details{
\code{eepd_sim()} calculates the absolute average prediction errors for each validation year and the ATT in the post-treatment year specified in the supplied \code{eepd_fits} object using each model fit to the data by \code{eepd_fit()}. These quantities can be computed directly from the model coefficients and the data. The optimal model is selected as that which minimizes the absolute average prediction error across validation years, and the reported ATT is the ATT computed using that model. \code{plot()} can be used to display the absolute average prediction errors across years for each model. The black bar represents the year with the largest absolute average prediction error for the corresponding model.

Because there is uncertainty in which model is best, simulation can be used to obtain a distribution of optimal models and their corresponding ATTs. New sets of parameters are drawn from a multivariate normal distribution, the assumed "posterior" distribution of the model coefficients, and in simulation, absolute average prediction errors are computed, the optimal model is chosen, and the ATT from that model is computed. For each simulation, this yields an optimal model and its ATT. The final reported ATT is the average of the simulated ATTs, which may be computed from different models. To request this simulation, \code{nsim} should be set to a large number (e.g., 1000). \code{plot()} can be used to display the distribution of the ATTs; the plot contains both the distribution of the ATTs by model and the distribution of optimal models. The vertical dotted line represents the ATT reported by \code{print()}, which is the average of the simulated ATTs from just the optimal models chosen in each simulation (i.e., \code{mean(out$atts)} where \code{out} is the output of \code{eepd_sim()} with \code{nsim > 0}).
}
\examples{
data("ptpdata")

# Combination of 8 models: 2 baseline formulas,
# 2 families, 2 lags
models <- eepd_mod(list(crude_rate ~ 1,
                        crude_rate ~ year),
                   family = list("gaussian", "quasipoisson"),
                   lag = 0:1, fixef = TRUE)
models

# Fit the models to data; unit_var must be supplied for
# fixed effects
fits <- eepd_fit(models, data = ptpdata,
                 group_var = "group",
                 time_var = "year",
                 val_times = 1999:2007,
                 post_time = 2008,
                 unit_var = "state")

# Choose the optimal model and compute its ATT
est <- eepd_sim(fits)

est

plot(est)

# Simulate average prediction errors and ATTs
cl <- parallel::detectCores()
est_sim <- eepd_sim(fits, nsim = 100,
                    cl = cl, verbose = TRUE)

est_sim

plot(est_sim)
}
\seealso{
\code{\link[=eepd_fit]{eepd_fit()}}; \code{\link[=eepd_boot]{eepd_boot()}} to bootstrap the simulation process to correctly account for uncertainty sampling as well as in choosing the optimal model.
}
